feat: all tests pass in ubuntu, other tests ignored with build tags
