+++ pkg/manager/decider_test.go
@@ -39,7 +39,7 @@ func TestDeciderMock(t *testing.T) {
 
 	t.Run("a failing skip_if runs", func(t *testing.T) {
 		r.RunFunc = func(ctx context.Context, printOnly bool, cmdLine string) (string, error) {
-			assert.Equal(t, "which apt", cmdLine)
+			assert.Equal(t, "which apk", cmdLine)
 			return "", errors.New("command exited with a non-zero exit code")
 		}
 		s := d.ShouldRun(context.Background(), []string{"which apk"}, nil)
@@ -64,42 +64,6 @@ func TestDeciderMock(t *testing.T) {
 			}
 			return "/usr/local/apt", errors.New("command exited with a non-zero exit code")
 		}
-		s := d.ShouldRun(context.Background(), []string{"which apk"}, []string{"which brew"})
-		assert.True(t, s)
-	})
-}
-
-//These tests are tightly coupled to my mac environment. It might be good to make a docker container as a test
-//environment and test there
-func TestDeciderMac(t *testing.T) {
-	d := NewDecider(io.NewShellRunner())
-
-	t.Run("empty should run_if and skip_if runs", func(t *testing.T) {
-		s := d.ShouldRun(context.Background(), nil, nil)
-		assert.True(t, s)
-	})
-
-	t.Run("passing run_if runs", func(t *testing.T) {
-		s := d.ShouldRun(context.Background(), nil, []string{"which brew"})
-		assert.True(t, s)
-	})
-
-	t.Run("a failing skip_if prohibits running", func(t *testing.T) {
-		s := d.ShouldRun(context.Background(), []string{"which brew"}, nil)
-		assert.False(t, s)
-	})
-
-	t.Run("passing skip_if runs", func(t *testing.T) {
-		s := d.ShouldRun(context.Background(), []string{"which apk"}, nil)
-		assert.True(t, s)
-	})
-
-	t.Run("passing run_if and failing skip_if prohibits running", func(t *testing.T) {
-		s := d.ShouldRun(context.Background(), []string{"which gvm"}, []string{"which brew"})
-		assert.False(t, s)
-	})
-
-	t.Run("passing run_if and passing skip_if runs", func(t *testing.T) {
 		s := d.ShouldRun(context.Background(), []string{"which apk"}, []string{"which brew"})
 		assert.True(t, s)
 	})
