+++ pkg/sync/sync_test.go
@@ -1,3 +1,5 @@
+//+build sync
+
 package sync
 
 import (
