+++ pkg/sync/walker_test.go
@@ -1,3 +1,6 @@
+//go:build sync
+// +build sync
+
 package sync
 
 import (
